+++ jib-core/src/main/java/com/google/cloud/tools/jib/frontend/JavaEntrypointConstructor.java
@@ -16,6 +16,7 @@
 
 package com.google.cloud.tools.jib.frontend;
 
+import com.google.common.base.Preconditions;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -23,17 +24,21 @@ import java.util.List;
 /** Constructs an image entrypoint for the Java application. */
 public class JavaEntrypointConstructor {
 
-  public static final String DEFAULT_RESOURCES_PATH_ON_IMAGE = "/app/resources/";
-  public static final String DEFAULT_CLASSES_PATH_ON_IMAGE = "/app/classes/";
-  public static final String DEFAULT_DEPENDENCIES_PATH_ON_IMAGE = "/app/libs/";
-  public static final String DEFAULT_JETTY_BASE_ON_IMAGE = "/jetty/webapps/ROOT/";
+  public static final String DEFAULT_RELATIVE_RESOURCES_PATH_ON_IMAGE = "resources/";
+  public static final String DEFAULT_RELATIVE_CLASSES_PATH_ON_IMAGE = "classes/";
+  public static final String DEFAULT_RELATIVE_DEPENDENCIES_PATH_ON_IMAGE = "libs/";
 
-  public static List<String> makeDefaultEntrypoint(List<String> jvmFlags, String mainClass) {
+  public static List<String> makeDefaultEntrypoint(
+      String appRoot, List<String> jvmFlags, String mainClass) {
+    Preconditions.checkArgument(
+        appRoot.startsWith("/"), "appRoot should be an absolute path in Unix-style: " + appRoot);
+    appRoot = appRoot.endsWith("/") ? appRoot : appRoot + '/';
+
     return makeEntrypoint(
         Arrays.asList(
-            DEFAULT_RESOURCES_PATH_ON_IMAGE,
-            DEFAULT_CLASSES_PATH_ON_IMAGE,
-            DEFAULT_DEPENDENCIES_PATH_ON_IMAGE + "*"),
+            appRoot + DEFAULT_RELATIVE_RESOURCES_PATH_ON_IMAGE,
+            appRoot + DEFAULT_RELATIVE_CLASSES_PATH_ON_IMAGE,
+            appRoot + DEFAULT_RELATIVE_DEPENDENCIES_PATH_ON_IMAGE + "*"),
         jvmFlags,
         mainClass);
   }