+++ jib-core/src/test/java/com/google/cloud/tools/jib/frontend/JavaLayerConfigurationsTest.java
@@ -2,7 +2,6 @@ package com.google.cloud.tools.jib.frontend;
 
 import com.google.cloud.tools.jib.configuration.LayerConfiguration;
 import com.google.cloud.tools.jib.image.LayerEntry;
-import com.google.common.collect.ImmutableList;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -17,48 +16,25 @@ import org.junit.Test;
 /** Tests for {@link JavaLayerConfigurations}. */
 public class JavaLayerConfigurationsTest {
 
+  private static JavaLayerConfigurations createFakeConfigurations() throws IOException {
+    return JavaLayerConfigurations.builder()
+        .setDependencyFiles(Collections.singletonList(Paths.get("dependency")), "/dependency/path")
+        .setSnapshotDependencyFiles(
+            Collections.singletonList(Paths.get("snapshot dependency")), "/snapshots")
+        .setResourceFiles(Collections.singletonList(Paths.get("resource")), "/resources/here")
+        .setClassFiles(Collections.singletonList(Paths.get("class")), "/classes/go/here")
+        .setExplodedWarFiles(Collections.singletonList(Paths.get("exploded war")), "/for/war")
+        .setExtraFiles(Collections.singletonList(Paths.get("extra file")), "/some/extras")
+        .build();
+  }
+
+  private static List<Path> layerEntriesToSourceFiles(List<LayerEntry> entries) {
+    return entries.stream().map(LayerEntry::getSourceFile).collect(Collectors.toList());
+  }
+
   @Test
-  public void testDefault() throws IOException {
-    JavaLayerConfigurations javaLayerConfigurations =
-        JavaLayerConfigurations.builder()
-            .setDependencyFiles(Collections.singletonList(Paths.get("dependency")))
-            .setSnapshotDependencyFiles(Collections.singletonList(Paths.get("snapshotDependency")))
-            .setResourceFiles(Collections.singletonList(Paths.get("resource")))
-            .setClassFiles(Collections.singletonList(Paths.get("class")))
-            .setExplodedWarFiles(Collections.singletonList(Paths.get("exploded-war")))
-            .setExtraFiles(Collections.singletonList(Paths.get("extra")))
-            .build();
-
-    ImmutableList<LayerEntry> dependencyLayerEntries =
-        javaLayerConfigurations.getDependencyLayerEntries();
-    ImmutableList<LayerEntry> snapshotDependencyLayerEntries =
-        javaLayerConfigurations.getSnapshotDependencyLayerEntries();
-    ImmutableList<LayerEntry> resourceLayerEntries =
-        javaLayerConfigurations.getResourceLayerEntries();
-    ImmutableList<LayerEntry> classLayerEntries = javaLayerConfigurations.getClassLayerEntries();
-    ImmutableList<LayerEntry> explodedWarLayerEntries =
-        javaLayerConfigurations.getExplodedWarEntries();
-    ImmutableList<LayerEntry> extraFilesLayerEntries =
-        javaLayerConfigurations.getExtraFilesLayerEntries();
-
-    Assert.assertEquals(
-        Paths.get(JavaEntrypointConstructor.DEFAULT_DEPENDENCIES_PATH_ON_IMAGE)
-            .resolve("dependency"),
-        dependencyLayerEntries.get(0).getExtractionPath());
-    Assert.assertEquals(
-        Paths.get(JavaEntrypointConstructor.DEFAULT_DEPENDENCIES_PATH_ON_IMAGE)
-            .resolve("snapshotDependency"),
-        snapshotDependencyLayerEntries.get(0).getExtractionPath());
-    Assert.assertEquals(
-        Paths.get(JavaEntrypointConstructor.DEFAULT_RESOURCES_PATH_ON_IMAGE).resolve("resource"),
-        resourceLayerEntries.get(0).getExtractionPath());
-    Assert.assertEquals(
-        Paths.get(JavaEntrypointConstructor.DEFAULT_CLASSES_PATH_ON_IMAGE).resolve("class"),
-        classLayerEntries.get(0).getExtractionPath());
-    Assert.assertEquals(
-        Paths.get(JavaEntrypointConstructor.DEFAULT_JETTY_BASE_ON_IMAGE).resolve("exploded-war"),
-        explodedWarLayerEntries.get(0).getExtractionPath());
-    Assert.assertEquals(Paths.get("/extra"), extraFilesLayerEntries.get(0).getExtractionPath());
+  public void testLabels() throws IOException {
+    JavaLayerConfigurations javaLayerConfigurations = createFakeConfigurations();
 
     List<String> expectedLabels = new ArrayList<>();
     for (JavaLayerConfigurations.LayerType layerType : JavaLayerConfigurations.LayerType.values()) {
@@ -72,42 +48,55 @@ public class JavaLayerConfigurationsTest {
   }
 
   @Test
-  public void testSetFiles() throws IOException {
-    List<Path> dependencyFiles = Collections.singletonList(Paths.get("/dependency"));
-    List<Path> snapshotDependencyFiles =
-        Collections.singletonList(Paths.get("/snapshot dependency"));
-    List<Path> resourceFiles = Collections.singletonList(Paths.get("/resource"));
-    List<Path> classFiles = Collections.singletonList(Paths.get("/class"));
-    List<Path> explodedWarFiles = Collections.singletonList(Paths.get("/exploded war"));
-    List<Path> extraFiles = Collections.singletonList(Paths.get("/extra file"));
-
-    JavaLayerConfigurations javaLayerConfigurations =
-        JavaLayerConfigurations.builder()
-            .setDependencyFiles(dependencyFiles)
-            .setSnapshotDependencyFiles(snapshotDependencyFiles)
-            .setResourceFiles(resourceFiles)
-            .setClassFiles(classFiles)
-            .setExplodedWarFiles(explodedWarFiles)
-            .setExtraFiles(extraFiles)
-            .build();
+  public void testSetFiles_files() throws IOException {
+    JavaLayerConfigurations javaLayerConfigurations = createFakeConfigurations();
 
     List<List<Path>> expectedFiles =
         Arrays.asList(
-            dependencyFiles,
-            snapshotDependencyFiles,
-            resourceFiles,
-            classFiles,
-            explodedWarFiles,
-            extraFiles);
-    List<List<Path>> actualFiles = new ArrayList<>();
-    for (LayerConfiguration layerConfiguration : javaLayerConfigurations.getLayerConfigurations()) {
-      actualFiles.add(
-          layerConfiguration
-              .getLayerEntries()
-              .stream()
-              .map(LayerEntry::getSourceFile)
-              .collect(Collectors.toList()));
-    }
+            Collections.singletonList(Paths.get("dependency")),
+            Collections.singletonList(Paths.get("snapshot dependency")),
+            Collections.singletonList(Paths.get("resource")),
+            Collections.singletonList(Paths.get("class")),
+            Collections.singletonList(Paths.get("exploded war")),
+            Collections.singletonList(Paths.get("extra file")));
+    List<List<Path>> actualFiles =
+        javaLayerConfigurations
+            .getLayerConfigurations()
+            .stream()
+            .map(LayerConfiguration::getLayerEntries)
+            .map(JavaLayerConfigurationsTest::layerEntriesToSourceFiles)
+            .collect(Collectors.toList());
     Assert.assertEquals(expectedFiles, actualFiles);
+  }
+
+  @Test
+  public void testSetFiles_extractionPaths() throws IOException {
+    JavaLayerConfigurations configurations = createFakeConfigurations();
+
+    Assert.assertEquals("/dependency/path", configurations.getDependencyExtractionPath());
+    Assert.assertEquals("/snapshots", configurations.getSnapshotDependencyExtractionPath());
+    Assert.assertEquals("/resources/here", configurations.getResourceExtractionPath());
+    Assert.assertEquals("/classes/go/here", configurations.getClassExtractionPath());
+    Assert.assertEquals("/for/war", configurations.getExplodedWarExtractionPath());
+    Assert.assertEquals("/some/extras", configurations.getExtraFilesExtractionPath());
+
+    Assert.assertEquals(
+        Paths.get("/dependency/path/dependency"),
+        configurations.getDependencyLayerEntries().get(0).getExtractionPath());
+    Assert.assertEquals(
+        Paths.get("/snapshots/snapshot dependency"),
+        configurations.getSnapshotDependencyLayerEntries().get(0).getExtractionPath());
+    Assert.assertEquals(
+        Paths.get("/resources/here/resource"),
+        configurations.getResourceLayerEntries().get(0).getExtractionPath());
+    Assert.assertEquals(
+        Paths.get("/classes/go/here/class"),
+        configurations.getClassLayerEntries().get(0).getExtractionPath());
+    Assert.assertEquals(
+        Paths.get("/for/war/exploded war"),
+        configurations.getExplodedWarEntries().get(0).getExtractionPath());
+    Assert.assertEquals(
+        Paths.get("/some/extras/extra file"),
+        configurations.getExtraFilesLayerEntries().get(0).getExtractionPath());
   }
 }